Add files via upload
You can either run the file directly in Visual Studio using the .sln file or copy the code file by file and then create your own project.

The code is written with comments to give the clarity what it really does. If something is not clear then you can ask me for it. 

Thanks,
Hitesh